[#28] docs : 패키지 변경
