+++ src/main/java/com/btb/chalKak/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package com.btb.chalKak.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
