+++ src/main/java/com/btb/chalKak/domain/hashTag/entity/HashTag.java
@@ -1,7 +1,7 @@
 package com.btb.chalKak.domain.hashTag.entity;
 
 import com.btb.chalKak.domain.post.entity.Post;
-import com.btb.chalKak.global.entity.BaseTimeEntity;
+import com.btb.chalKak.common.entity.BaseTimeEntity;
 import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
