+++ src/main/java/com/btb/chalKak/domain/styleTag/entity/StyleTag.java
@@ -3,7 +3,7 @@ package com.btb.chalKak.domain.styleTag.entity;
 import com.btb.chalKak.domain.member.entity.Member;
 import com.btb.chalKak.domain.post.entity.Post;
 import com.btb.chalKak.domain.styleTag.type.StyleCategory;
-import com.btb.chalKak.global.entity.BaseTimeEntity;
+import com.btb.chalKak.common.entity.BaseTimeEntity;
 import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
