--- src/main/java/com/btb/chalKak/global/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.btb.chalKak.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
