--- src/main/java/com/btb/chalKak/global/mapper/EntityMapper.java
@@ -1,14 +1,0 @@
-package com.btb.chalKak.global.mapper;
-
-import java.util.List;
-
-public interface EntityMapper<D, E> {
-
-    E toEntity(final D dto);
-
-    D toDto(final E entity);
-
-    List<D> toDtoList(List<E> entityList);
-
-    List<E> toEntityList(List<D> dtoList);
-}
