--- src/main/java/com/btb/chalKak/global/oauth2/handler/OAuth2SuccessHandler.java
@@ -1,23 +1,0 @@
-package com.btb.chalKak.global.oauth2.handler;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@RequiredArgsConstructor
-@Component
-public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
-
-  @Override
-  public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
-      Authentication authentication) throws IOException, ServletException {
-      log.info("OAuth2 Login 성공!");
-  }
-}
